requirements.txt
elasticsearch==6.1.1
pika==0.11.2
requests==2.18.4
urllib3==1.22
logging==0.4.9.6
python-logstash==0.4.6